Use refactored code instead of original.
